# Implement a key-based process table