Processes are now uniquely identified by a ProcessKey instead of
process IDs as those may be recycled quickly.

Most of the code interfacing with the process table has been pulled
into the early part of Coalesce::transform_event that handles SYSCALL
messages.

The test introduced with #168 now passes.

Fixes #146

Performance-wise, this is a slight improvement over current master.

This branch:

```
test parse_only      ... bench:  20,237,394 ns/iter (+/- 235,818)
test parse_serialize ... bench:  27,426,770 ns/iter (+/- 255,343)
```

master (f237694c3cea11ab805ea2ebf4058a3106018821):

```
test parse_only      ... bench:  22,737,572 ns/iter (+/- 397,532)
test parse_serialize ... bench:  30,010,731 ns/iter (+/- 369,789)
```